SIGARNV-3335: Modifico los métodos y sus respectivos updates de las operaciones de Anular, validar, denegar y eliminar, para ello que tenido que refrescar los mappers usando myBatis para que reconozcan la nueva columna ELIMINADO de BBDD en la tabla CEN_BAJASTEMPORALES, ya que se creó pero no se actualizaron sus mappers en back.

+++ siga-web/src/main/java/org/itcgae/siga/scs/services/impl/guardia/CargasMasivasGuardiaServiceImpl.java
@@ -2040,7 +2040,7 @@ public class CargasMasivasGuardiaServiceImpl implements CargasMasivasGuardiaServ
 						cenBajasTemporales.setFechabt(new Date());
 						cenBajasTemporales.setDescripcion(cargaMasivaDatosBTItem.getMotivo());
 						cenBajasTemporales.setFechaalta(new Date());
-						cenBajasTemporales.setEliminado(0);
+						cenBajasTemporales.setEliminado((short) 0);
 						cenBajasTemporales.setValidado("1");
 
 						// Comprobar que las fechas introducidas en el fichero están dentro del periodo
